requirements.txt
asgiref==3.2.10
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap3==14.1.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
